Added code and example data to project
Added the following to the project:
Code to calculate a ARIMA model to the time series data, using some
fraction as training and the remainder as test data, then calculate a
prediction/forecast on the test data on a per-datapoint basis. Added
sample shampoo sale data, and a predicitons output file which the code
writes output to.